View recycling for attachments
- Preallocates and re-uses attachment views inside the attachments view
- Removes Equatable and comparisons for attachment view models (introduced in 718b2c2d23fbb8d7fa459c724046b98333aabebd)
- Fixes #99, which was introduced by that change
- Fixes the animation bug that the Equatable part of 718b2c2d23fbb8d7fa459c724046b98333aabebd was trying to fix

The tl;dr of this bug is that adding or removing subviews from a view model `didSet` on a view almost always causes animation glitches with table cells that contain that view.

The cause appears to be that the view model is set repeatedly during the slide animation when a timeline post or notification is focused and presented, causing the subviews to be changed repeatedly, which in this case caused the attachment views to appear as if they were being expanded from their lower left corner instead of smoothly sliding with the rest of the status view. (This isn't an animation we control directly; I'm not sure which UIKit component is responsible for it.)

In some cases, we can fix this by de-duping the view model changes: in `didSet`, check to see if the view model is the same as the currently set one, and if so, do nothing. Unfortunately, the view model setters in this codebase aren't idempotent, and a lot of code relies on the side effects of setting a view model, including, apparently, the part that lets you edit image descriptions.

So to handle this case, we avoid changing the attachment view subviews of the attachment view by pre-allocating all 4 of them and inserting them into the view hierarchy at attachments view initialization, then showing and hiding them as required, which does not trigger the animation glitch. This assumes that the number of media attachments is capped at 4, which is not the case for all Fedi servers, but I'm not going to address that today.