Fix listing blocked users on GotoSocial
GtS considers an account's avatar, header, emojis, and fields to be optional, and when reading a blocked account, will send empty strings for avatar and header URLs and null instead of an empty list for emojis and fields. This commit is mostly the consequences of making the UnicodeURL wrapper partially fallible in case the contents of the field are not a valid URL, but also sets default empty lists for emojis and fields. (I decided to make UnicodeURL allow nil URLs, rather than making some or all UnicodeURL properties optional, because it involved no content or identity DB changes.)